First implementation of the LOS effects in lenstronomy
Line-of-sight corrections are implemented so that lenstronomy understands them as an additional lens. The associated 'fake profile' is called LOS in los.py.
This is then interpreted by the class SinglePlaneLOS(). Since I modified the file a lot, I have put it in a new file called single_plane_los_2.