Correcting the ordering bug
There was a bug in the lens equation solver when we were adding a LOS profile in first position of the lens profile list. This revealed a deeper problem coming from the omission of a proper __init__ in the class SinglePlaneLOS, which was not affected a func_list.
I hope that correcting this did not break it all...