Add primary beam to image simulation (ImageModel) and linear solver (ImageLinearFit)
Add the primary beam to image simulation and linear solver (where the light response is computed).

The 'primary beam' is input into ImageData class and is inherited by ImageModel and further ImageLinearFit.
It should be a 2d numpy array and has the same size of the data image.

Physically, the primary beam is the angular response function of each antenna of an interferometry array. Numerically, it serves as a multiplier on an unconvolved image.

In other words, a convolved interferometric image is
I_convolved = (I_unconvolved . primary beam ) * PSF,
where . denotes multiplication and * denotes convolution.

For more info on primary beam, consult
https://casaguides.nrao.edu/index.php/VLA_CASA_Imaging-CASA5.0.0#Primary_and_Synthesized_Beam
https://casadocs.readthedocs.io/en/stable/notebooks/synthesis_imaging.html?highlight=primary%20beam%20#Antenna-Voltage/Power-Patterns:-Primary-Beam